simple and multiple compilation examples

+++ multiple/kernel.cu
@@ -1,4 +1,4 @@
-__global__ void kernel(int *a) {
+__global__ void kernel(float *a) {
     int i = gridDim.x;
     a[i] = i;
 }
